+++ changelog.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+ - The functions OpenEventLog, CloseEventLog and ReadEventLog were added to
+   advapi.
+ - The sample uptime_evt.py demonstrates a way how these functions can be used.
+
 ## [1.0.0]
 
 ### Fixed